Created a Dockerfile. (#2)
Commits:

C1. Created a Dockerfile.
Created a Dockerfile to defined the environment in which the action will run.

C2. Created an entrypoint.sh file, which defines what the action will do.
This entrypoint.sh file is executed by Docker.

C3. Created an action.yaml file, which defines metadata for the action.

C4. Created a main.yml file to kickstart the workflow file for the action.

C5. Appended a job to run the action to the main.yml workflow file.