Tweaked the action's echo message. (#4)
Commits:

C1. Switched to an Australian greeting for the action.

C2. Attempt 1 to fix syntax issue with the action's Aussie echo message.
Enclosed the variable name in curly braces in the hope that it separates the question mark at the end of the greeting from the variable name.

C3. Attempt 2: Fixed the action's echo message.
There was no need to wrap the variable name in curly braces; I simply needed to escape the quotation mark in "goin'".